pyproject.toml
[project]
name = "demopelican"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "invoke>=2.2.0",
    "livereload>=2.7.1",
    "pelican[markdown]>=4.11.0.post0",
    "pillow>=10.4.0",
]
